renamed utils layer to shared.
