+++ lib/bootstrap.dart
@@ -4,7 +4,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:groceries/config/injection/injector.dart';
-import 'package:groceries/utils/logger.dart';
+import 'package:groceries/shared/logger.dart';
 import 'package:talker_bloc_logger/talker_bloc_logger.dart';
 
 /// Bootstraps the Flutter application by setting up error handling and
