+++ lib/data/repositories/priorities_repository_impl.dart
@@ -3,7 +3,7 @@ import 'package:groceries/data/data_sources/local/cache_priorities_data_source.d
 import 'package:groceries/data/data_sources/remote/firestore_priorities_data_source.dart';
 import 'package:groceries/data/models/priority/priority_model.dart';
 import 'package:groceries/domain/repositories/priorities_repository.dart';
-import 'package:groceries/utils/logger.dart';
+import 'package:groceries/shared/logger.dart';
 import 'package:injectable/injectable.dart';
 
 @Singleton(as: PrioritiesRepository)
