+++ lib/presentation/blocs/grocery_lists/grocery_lists_bloc.dart
@@ -4,7 +4,7 @@ import 'package:groceries/config/localization/translations.dart';
 import 'package:groceries/domain/entities/grocery_list/grocery_list_entity.dart';
 import 'package:groceries/domain/use_cases/grocery_lists_use_case.dart';
 import 'package:groceries/presentation/common/bloc_status.dart';
-import 'package:groceries/utils/exceptions/app_network_exception.dart';
+import 'package:groceries/shared/exceptions/app_network_exception.dart';
 import 'package:injectable/injectable.dart';
 
 part 'grocery_lists_bloc.freezed.dart';
