+++ lib/presentation/modules/forgot_password/bloc/forgot_password_bloc.dart
@@ -4,7 +4,7 @@ import 'package:formz/formz.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:groceries/domain/use_cases/authentication_use_cases/send_password_reset_email_use_case.dart';
 import 'package:groceries/presentation/forms/email_form.dart';
-import 'package:groceries/utils/exceptions/app_network_exception.dart';
+import 'package:groceries/shared/exceptions/app_network_exception.dart';
 import 'package:injectable/injectable.dart';
 
 part 'forgot_password_bloc.freezed.dart';
