+++ lib/presentation/modules/grocery_list_details/bloc/grocery_list_details_bloc.dart
@@ -7,7 +7,7 @@ import 'package:groceries/domain/use_cases/clear_grocery_list_items_use_case.dar
 import 'package:groceries/domain/use_cases/fetch_grocery_list_details_use_case.dart';
 import 'package:groceries/domain/use_cases/remove_item_from_list_use_case.dart';
 import 'package:groceries/presentation/common/bloc_status.dart';
-import 'package:groceries/utils/exceptions/app_network_exception.dart';
+import 'package:groceries/shared/exceptions/app_network_exception.dart';
 import 'package:injectable/injectable.dart';
 
 part 'grocery_list_details_bloc.freezed.dart';
