+++ lib/presentation/modules/grocery_list_settings/blocs/delete_list/delete_list_bloc.dart
@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:groceries/config/localization/translations.dart';
 import 'package:groceries/domain/use_cases/delete_grocery_list_use_case.dart';
 import 'package:groceries/presentation/common/bloc_status.dart';
-import 'package:groceries/utils/exceptions/app_network_exception.dart';
+import 'package:groceries/shared/exceptions/app_network_exception.dart';
 
 part 'delete_list_bloc.freezed.dart';
 part 'delete_list_event.dart';
