+++ lib/presentation/modules/update_display_name/bloc/update_display_name_bloc.dart
@@ -4,7 +4,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:groceries/config/localization/translations.dart';
 import 'package:groceries/domain/use_cases/authentication_use_cases/update_display_name_use_case.dart';
 import 'package:groceries/presentation/forms/display_name_form.dart';
-import 'package:groceries/utils/exceptions/app_network_exception.dart';
+import 'package:groceries/shared/exceptions/app_network_exception.dart';
 import 'package:injectable/injectable.dart';
 
 part 'update_display_name_bloc.freezed.dart';
