+++ lib/shared/exceptions/app_network_exception.dart
